+++ passes/V004/V004_test.go
@@ -1,0 +1,13 @@
+package V004_test
+
+import (
+	"testing"
+
+	"github.com/bflad/tfproviderlint/passes/V004"
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func TestV004(t *testing.T) {
+	testdata := analysistest.TestData()
+	analysistest.Run(t, testdata, V004.Analyzer, "a")
+}