+++ passes/V005/V005_test.go
@@ -1,0 +1,13 @@
+package V005_test
+
+import (
+	"testing"
+
+	"github.com/bflad/tfproviderlint/passes/V005"
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func TestV005(t *testing.T) {
+	testdata := analysistest.TestData()
+	analysistest.Run(t, testdata, V005.Analyzer, "a")
+}