+++ passes/V006/V006_test.go
@@ -1,0 +1,13 @@
+package V006_test
+
+import (
+	"testing"
+
+	"github.com/bflad/tfproviderlint/passes/V006"
+	"golang.org/x/tools/go/analysis/analysistest"
+)
+
+func TestV006(t *testing.T) {
+	testdata := analysistest.TestData()
+	analysistest.Run(t, testdata, V006.Analyzer, "a")
+}