+++ passes/V006/testdata/src/a/main.go
@@ -1,0 +1,22 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform-plugin-sdk/helper/validation"
+)
+
+func f() {
+	// Passing
+
+	_ = validation.ListOfUniqueStrings
+
+	// Comment ignored
+
+	//lintignore:V006
+	_ = validation.ValidateListUniqueStrings
+
+	_ = validation.ValidateListUniqueStrings //lintignore:V006
+
+	// Failing
+
+	_ = validation.ValidateListUniqueStrings // want "deprecated validation.ValidateListUniqueStrings should be replaced with validation.ListOfUniqueStrings"
+}