+++ passes/selectorexpr/helper/validation/singleip/singleip.go
@@ -1,0 +1,13 @@
+package singleip
+
+import (
+	"github.com/bflad/tfproviderlint/helper/analysisutils"
+	"github.com/bflad/tfproviderlint/helper/terraformtype/helper/validation"
+)
+
+var Analyzer = analysisutils.SelectorExprAnalyzer(
+	"singleip",
+	validation.IsPackageFunc,
+	validation.PackagePath,
+	validation.FuncNameSingleIP,
+)